Add files via upload
The car dealership program is written in Python. It makes a mini bank account with a set balance, and alters the balance based on what car from the dictionary of cars you buy. Each car is assigned a certain value (the price) and this value decrements the balance if the user could afford the car. The class "account" tells the user if you can afford the car or not depending on what you select from the choices.