Add files via upload
Unfinished project due to lack of research and development in natural language processing. I tried to make a small app that received the premises and conclusion of an argument, and it would check to see if the overall argument was valid based on if the conclusion was true. I tried looking up ways for the computer to read human input and determine if a sentence is true or false, but I couldn't find any possible solutions yet. Definitely a fun little project though.